Extract weather logic to dedicated module
