+++ js/simulation.js
@@ -1,4 +1,4 @@
-import { clamp, clamp01, lerp, log, randomNormal } from './utils.js';
+import { clamp, clamp01, lerp, log } from './utils.js';
 import {
   CROPS,
   WORK_MINUTES,
@@ -35,6 +35,7 @@ import { attachPastureIfNeeded } from './world.js';
 import { rowGrowthMultiplier } from './state.js';
 import { autosave } from './persistence.js';
 import { MINUTES_PER_DAY, computeDaylightByIndex, dayIndex } from './time.js';
+import { generateWeatherToday, dailyWeatherEvents } from './weather.js';
 
 export { processFarmerHalfStep } from './tasks.js';
 import { assertNoWorkOutsideWindow } from './tests/invariants.js';
@@ -281,22 +282,6 @@ export function consumeLivestock(world) {
   if (S.hay < 1) world.alerts.push('Hay low');
 }
 
-export function generateWeatherToday(world) {
-  const m = world.calendar.month;
-  const rng = world.rng;
-  const base = WX_BASE[m];
-  const temp = base.tMean + 3.0 * randomNormal(rng);
-  const wetChance = 0.45 + (base.rainMean - 2.0) * 0.06;
-  const rain = (rng() < wetChance) ? Math.max(0, base.rainMean + 5 * randomNormal(rng)) : 0;
-  const wind = Math.max(0, 2 + 2 * randomNormal(rng));
-  const frost = (m <= 2) && (temp < 2) && (rng() < 0.3);
-  world.weather.tempC = temp;
-  world.weather.rain_mm = Math.max(0, rain);
-  world.weather.wind_ms = wind;
-  world.weather.frostTonight = !!frost;
-  world.weather.dryStreakDays = (rain <= 0.2) ? (world.weather.dryStreakDays + 1) : 0;
-}
-
 export function updateSoilWaterDaily(world) {
   const W = world.weather;
   const rain = W.rain_mm;
@@ -329,30 +314,6 @@ export function updateHayCuring(world) {
       h.dryness = Math.max(0, h.dryness - 0.15);
       h.loss_t += Math.min(h.mass_t * 0.03, 0.1);
     }
-  }
-}
-
-export function dailyWeatherEvents(world) {
-  const w = world.weather;
-  const m = world.calendar.month;
-  if (w.frostTonight) {
-    const g = world.parcels[world.parcelByKey.homestead];
-    g.status.frost = (g.status.frost || 0) + 1;
-    const o = world.parcels[world.parcelByKey.orchard];
-    o.status.frostBites = (o.status.frostBites || 0) + 1;
-  }
-  if (m >= 3 && m <= 5 && w.wind_ms >= 10) {
-    const hit = [];
-    for (const key of ['barley_clover', 'oats_close', 'pulses', 'flex', 'wheat']) {
-      const p = world.parcels[world.parcelByKey[key]];
-      if (!p || !p.rows?.length) continue;
-      const matureish = p.rows.some(r => r.crop && r.growth > 0.6);
-      if (matureish && (p.status.mud || 0) > 0.2) {
-        p.status.lodgingPenalty = Math.max(p.status.lodgingPenalty || 0, 0.08 + 0.04 * Math.random());
-        hit.push(p.name);
-      }
-    }
-    if (hit.length) (world.alerts = world.alerts || []).push(`Storm lodging: ${hit.join(', ')}`);
   }
 }
